+++ include/mathfunc_parallel_cu.h
@@ -13,5 +13,5 @@ void reduce_sum2_cu(HT (&h_ans)[HN], DPTR v, size_t nelem, int queue);
 
 
 template <class T>
-void dotprod_cu(T* ans, const T* a, const T* b, int nelem, int queue);
+void dotprod_cu(T* ans, const T* a, const T* b, size_t nelem, int queue);
 }
